feat: :sparkles: 3) Una tarea
Al hacer doble click se puede editar la tarea selecionada

Se actualizó la versión de Playwright para poder usar la extensión de testing de VSCode.
