+++ .gitignore
@@ -105,3 +105,7 @@ dist
 /test-results/
 /playwright-report/
 /playwright/.cache/
+/test-results/
+/playwright-report/
+/blob-report/
+/playwright/.cache/
