+++ package.json
@@ -23,7 +23,8 @@
 	"devDependencies": {
 		"@babel/core": "7.18.6",
 		"@babel/preset-env": "7.18.6",
-		"@playwright/test": "1.23.0",
+		"@playwright/test": "^1.28.0",
+		"@types/node": "^20.10.5",
 		"@typescript-eslint/eslint-plugin": "^6.16.0",
 		"@typescript-eslint/parser": "^6.16.0",
 		"babel-loader": "8.2.5",
