+++ src/index.html
@@ -45,7 +45,7 @@
 					<ul class="todo-list">
 						<!-- These are here just to show the structure of the list items -->
 						<!-- List items should get the class `editing` when editing and `completed` when marked as completed -->
-						<li class="completed">
+						<!-- <li class="completed">
 							<div class="view">
 								<input
 									class="toggle"
@@ -87,7 +87,7 @@
 								class="edit"
 								value="Make dishes"
 							/>
-						</li>
+						</li> -->
 					</ul>
 				</section>
 				<!-- This footer should be hidden by default and shown when there are todos -->
